requirements.txt
playwright>=1.40.0
requests>=2.31.0
openai>=1.12.0
qrcode[pil]>=7.4
schedule>=1.2.0
python-dotenv>=1.0.0
Pillow>=10.0.0
